feat(bot): initial version of bot
Bot currently supports starting a game with the start command which will
hook up the SDK into AmongUs. Once that has happend, events are
subscribed to for actions such as joining a server, game state change.

Right now the discrod bot will post a message when a lobby is joined
which includes the access code and region. The other feature is an auto
mute system in which the bot will automatically mute players during play
time and unmute alive players during voting/disucussion points in the
game. Bot is still very rough around the edges and a lot of improvements
can be made.